+++ src/components/SQForm/SQFormAutocomplete.tsx
@@ -77,12 +77,12 @@ const OuterElementType = React.forwardRef<HTMLDivElement>(
   (props: OuterElementTypeProps, ref) => {
     const outerProps = React.useContext(OuterElementContext);
     return (
-      <div
+      <Box
         ref={ref}
         {...props}
         {...outerProps}
         className={`${props?.className || ''} ${outerProps?.className || ''}`}
-        style={{...props?.style, ...outerProps?.style}}
+        sx={{...[props?.style || {}], ...[outerProps?.style || {}]}}
       />
     );
   }