+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [10.2.0](https://github.com/SelectQuoteLabs/SQForm/compare/v10.1.0...v10.2.0) (2023-02-07)
+
+
+### Features
+
+* 🎸 Upgrade to SCPlus-Shared-Components@12 ([7c40272](https://github.com/SelectQuoteLabs/SQForm/commit/7c402728a000e7f4ce8101deb89b89185eb565c8)), closes [#828](https://github.com/SelectQuoteLabs/SQForm/issues/828)
+
+
+### Bug Fixes
+
+* 🐛 Fix SQFormDatePicker MUI Utils issue ([2ceb6ee](https://github.com/SelectQuoteLabs/SQForm/commit/2ceb6ee06732e08f094ca893b718acd5b668546a))
+
+## [10.1.0](https://github.com/SelectQuoteLabs/SQForm/compare/v10.0.2...v10.1.0) (2023-02-07)
+
+
+### Features
+
+* 🎸 Upgrade to SCPlus-Shared-Components@12 ([737f2f1](https://github.com/SelectQuoteLabs/SQForm/commit/737f2f15e53d65530b197285bb898b7514f525dc)), closes [#828](https://github.com/SelectQuoteLabs/SQForm/issues/828)
+
 ### [10.0.2](https://github.com/SelectQuoteLabs/SQForm/compare/v10.0.1...v10.0.2) (2023-01-20)